+++ src/test/java/lotto/domain/vo/ProfitTest.java
@@ -1,10 +1,8 @@
-package lotto.domain;
+package lotto.domain.vo;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import lotto.domain.vo.Money;
-import lotto.domain.vo.Profit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
